+++ ha-auto-guest-login/run.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/with-contenv bashio
-
-cd src;
-node server.js;