+++ ha-auto-guest-login/src/utils.js
@@ -1,9 +1,0 @@
-export class Utils {
-  constructor(addonPort) {
-    this.addonPort = addonPort;
-  }
-
-  getGuestLoginUrl(req) {
-    return `${req.protocol}://${req.hostname}:${this.addonPort}`;
-  }
-}